parser: rewrite binexpToExp.
The issue was in the end not in our original implementation of the
shunting yard algorithm, that was producing a correct postfix expression
chain.

The real issue was in the method we were using to translate from postfix
to a BinOp tree. The original algorithm handled associativity when
translating, it didn't need tweaking at that point.

At any rate, I feel the all in one algorithm is cleaner, and it's at least
_more_ correct than the previous implementation.