semantic: Improve type lookup error messages.
Previously, a failed lookup in a scope would always illicit

    SCOPE doesn't have member TYPENAME

Which is silly right out of the gate, as we don't know the
user thought it was a member.

So now single identifier (or the first in a chain) will result
in a simple unknown identifier error, and in other cases the
actual failed lookup will be denoted.